folder struct +osx fix
