+++ src/DurableOrchestrator/Models/BlobStorageInfo.cs
@@ -1,0 +1,18 @@
+namespace DurableOrchestrator.Models;
+public class BlobStorageInfo
+{
+    [JsonPropertyName("storageAccountName")]
+    public string StorageAccountName { get; set; } = string.Empty;
+    [JsonPropertyName("blobName")]
+    public string BlobName { get; set; } = string.Empty;
+
+    [JsonPropertyName("containerName")]
+    public string ContainerName { get; set; } = string.Empty;
+
+    [JsonPropertyName("blobUri")]
+    public string BlobUri { get; set; } = string.Empty;
+    [JsonPropertyName("content")]
+    public string Content { get; set; } = string.Empty;
+    [JsonPropertyName("buffer")]
+    public byte[] Buffer { get; set; } = Array.Empty<byte>();
+}
