+++ src/DurableOrchestrator/Models/ValidationResult.cs
@@ -1,0 +1,13 @@
+namespace DurableOrchestrator.Models;
+
+public class ValidationResult
+{
+    public bool IsValid { get; set; }
+    public List<string> Errors { get; set; } = new List<string>();
+
+    public void AddError(string error)
+    {
+        IsValid = false;
+        Errors.Add(error);
+    }
+}
