+++ src/DurableOrchestrator/Program.cs
@@ -31,4 +31,3 @@ host.Run();
 /// </summary>
 public partial class Program;
 
-
